requirements.txt
colorzero==2.0
gpiozero==1.6.2
mpmath==1.3.0
numpy==1.26.4
pydub==0.25.1
pygame==2.5.2
RPi.GPIO==0.7.0
scipy==1.12.0
six==1.16.0
sympy==1.12
toml==0.10.2
transitions==0.8.10
